+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules.hs
@@ -13,6 +13,7 @@ module Cardano.Ledger.Dijkstra.Rules (
   module Cardano.Ledger.Dijkstra.Rules.SubCert,
   module Cardano.Ledger.Dijkstra.Rules.SubCerts,
   module Cardano.Ledger.Dijkstra.Rules.SubDeleg,
+  module Cardano.Ledger.Dijkstra.Rules.SubEntities,
   module Cardano.Ledger.Dijkstra.Rules.SubGovCert,
   module Cardano.Ledger.Dijkstra.Rules.SubGov,
   module Cardano.Ledger.Dijkstra.Rules.SubLedger,
@@ -42,6 +43,7 @@ import Cardano.Ledger.Dijkstra.Rules.Pool ()
 import Cardano.Ledger.Dijkstra.Rules.SubCert
 import Cardano.Ledger.Dijkstra.Rules.SubCerts
 import Cardano.Ledger.Dijkstra.Rules.SubDeleg
+import Cardano.Ledger.Dijkstra.Rules.SubEntities
 import Cardano.Ledger.Dijkstra.Rules.SubGov
 import Cardano.Ledger.Dijkstra.Rules.SubGovCert
 import Cardano.Ledger.Dijkstra.Rules.SubLedger